requirements.txt
blinker==1.8.1
click==8.1.7
coverage==7.5.3
dnspython==2.6.1
email_validator==2.1.1
Flask==3.0.3
Flask-Mail==0.9.1
Flask-SQLAlchemy==3.1.1
Flask-WTF==1.2.1
greenlet==3.0.3
idna==3.7
iniconfig==2.0.0
install==1.3.5
itsdangerous==2.2.0
Jinja2==3.1.3
MarkupSafe==2.1.5
packaging==24.1
pluggy==1.5.0
pytest==8.2.2
pytest-cov==5.0.0
pytest-mock==3.14.0
python-dotenv==1.0.1
SQLAlchemy==2.0.30
typing_extensions==4.11.0
Werkzeug==3.0.2
WTForms==3.1.2
